Remove Last Legacy Code
Rescue the page's tests, and drop the remaining items in `legacy/`.
